@@ -1,67 +1,100 @@
 library ads;
 
-///
-/// Copyright (C) 2019 Andrious Solutions
-///
-/// Licensed under the Apache License, Version 2.0 (the "License");
-/// you may not use this file except in compliance with the License.
-/// You may obtain a copy of the License at
-///
-///    http://www.apache.org/licenses/LICENSE-2.0
-///
-/// Unless required by applicable law or agreed to in writing, software
-/// distributed under the License is distributed on an "AS IS" BASIS,
-/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-/// See the License for the specific language governing permissions and
-/// limitations under the License.
-///
-///          Created  04 Aug 2019
-///
-///
-
-import 'dart:io' show Platform;
-
-import 'package:flutter/widgets.dart' show State;
+// Copyright 2019 Andrious Solutions Ltd. All rights reserved.
+// Use of this source code is governed by the Apache License, Version 2.0 that can be
+// found in the LICENSE file.
 
 import 'dart:async' show Future;
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart' show Container, Widget;
 
-import 'package:firebase_admob/firebase_admob.dart';
+import 'package:google_mobile_ads/google_mobile_ads.dart';
 
-/// Signature for a [RewardAd] status change callback.
-typedef void RewardListener(String rewardType, int rewardAmount);
+// import 'package:firebase_admob/firebase_admob.dart'
+//     hide AdSize, BannerAd, InterstitialAd, MobileAdListener, NativeAd;
 
-/// Signature for a [AdError] status change callback.
-typedef void AdErrorListener(Exception ex);
+/// The Ads event function signature
+typedef AdsListener = void Function(Ad ad);
+
+/// Signature for a ['MobileAd'] status change callback.
+typedef MobileAdListener = void Function(AdsEvent event);
+
+/// Signature for a ['RewardAd'] status change callback.
+typedef RewardListener = void Function(String rewardType, int rewardAmount);
+
+/// Signature for a ['AdError'] status change callback.
+typedef AdErrorListener = void Function(Exception ex);
 
 class Banner extends MobileAds {
   //
-  factory Banner({MobileAdListener listener}) => _this ??= Banner._(listener);
+  factory Banner({required AdMobListener listener}) =>
+      _this ??= Banner._(listener);
+  Banner._(AdMobListener listener) : super(listener: listener) {
+    listener.closedListener = () {
+      loadAd(show: false);
+    };
+  }
+  static Banner? _this;
+  AdSize? _setSize;
+  AdSize? _showSize;
 
-  Banner._(MobileAdListener listener) : super(listener: listener);
+//  final _bannerListener = AdMobListener(adEventListeners);
 
-  static Banner _this;
-
-  AdSize _setSize;
-  AdSize _showSize;
+  Set<MobileAdListener> get eventListeners => listener!.eventListeners;
+  // ignore: avoid_setters_without_getters
+  set loadedListener(VoidCallback listener) =>
+      this.listener!.loadedListener = listener;
+  bool removeLoaded(VoidCallback listener) =>
+      this.listener!.removeLoaded(listener);
+  // ignore: avoid_setters_without_getters
+  set failedListener(VoidCallback listener) =>
+      this.listener!.failedListener = listener;
+  bool removeFailed(VoidCallback listener) =>
+      this.listener!.removeFailed(listener);
+  // ignore: avoid_setters_without_getters
+  set clickedListener(VoidCallback listener) =>
+      this.listener!.clickedListener = listener;
+  bool removeClicked(VoidCallback listener) =>
+      this.listener!.removeClicked(listener);
+  // ignore: avoid_setters_without_getters
+  set impressionListener(VoidCallback listener) =>
+      this.listener!.impressionListener = listener;
+  bool removeImpression(VoidCallback listener) =>
+      this.listener!.removeImpression(listener);
+  // ignore: avoid_setters_without_getters
+  set openedListener(VoidCallback listener) =>
+      this.listener!.openedListener = listener;
+  bool removeOpened(VoidCallback listener) =>
+      this.listener!.removeOpened(listener);
+  // ignore: avoid_setters_without_getters
+  set leftAppListener(VoidCallback listener) =>
+      this.listener!.leftAppListener = listener;
+  bool removeLeftApp(VoidCallback listener) =>
+      this.listener!.removeLeftApp(listener);
+  // ignore: avoid_setters_without_getters
+  set closedListener(VoidCallback listener) =>
+      this.listener!.closedListener = listener;
+  bool removeClosed(VoidCallback listener) =>
+      this.listener!.removeClosed(listener);
+  void clearAll() => listener!.clearAll();
 
   /// Set options to the Banner Ad.
   @override
   Future<bool> set({
-    String adUnitId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool nonPersonalizedAds,
-    bool testing,
-    AdSize size,
-    double anchorOffset,
-    double horizontalCenterOffset,
-    AnchorType anchorType,
-    AdErrorListener errorListener,
+    String? adUnitId,
+    AdRequest? targetInfo,
+    List<String>? keywords,
+    String? contentUrl,
+    bool? childDirected,
+    List<String>? testDevices,
+    bool? nonPersonalizedAds,
+    bool? testing,
+    AdSize? size,
+    double? anchorOffset,
+    double? horizontalCenterOffset,
+//    AnchorType anchorType,
+    AdErrorListener? errorListener,
   }) {
     super.set(
       adUnitId: adUnitId,
@@ -74,11 +107,11 @@ class Banner extends MobileAds {
       testing: testing,
       anchorOffset: anchorOffset,
       horizontalCenterOffset: horizontalCenterOffset,
-      anchorType: anchorType,
+//      anchorType: anchorType,
       errorListener: errorListener,
     );
     _setSize ??= size ?? AdSize.banner;
-    return load(
+    return loadAd(
       show: false,
       adUnitId: adUnitId,
       targetInfo: targetInfo,
@@ -90,81 +123,96 @@ class Banner extends MobileAds {
       testing: testing,
       anchorOffset: anchorOffset,
       horizontalCenterOffset: horizontalCenterOffset,
-      anchorType: anchorType,
+//      anchorType: anchorType,
     );
   }
 
-  /// Show the Banner Ad.
-  @override
-  Future<bool> show({
-    String adUnitId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool testing,
-    AdSize size,
-    double anchorOffset,
-    double horizontalCenterOffset,
-    AnchorType anchorType,
-    State state,
-  }) async {
-    if (size != null) dispose();
-    _showSize = size ?? _setSize ?? AdSize.banner;
-    return super.show(
-        adUnitId: adUnitId,
-        targetInfo: targetInfo,
-        keywords: keywords,
-        contentUrl: contentUrl,
-        childDirected: childDirected,
-        testDevices: testDevices,
-        testing: testing,
-        anchorOffset: anchorOffset,
-        horizontalCenterOffset: horizontalCenterOffset,
-        anchorType: anchorType,
-        state: state);
-  }
+  /// Return the Banner Ad.
+  Widget get widget => adWidget();
 
   @override
-  _createAd({bool testing, String adUnitId, MobileAdTargetingInfo targetInfo}) {
-    return BannerAd(
-      adUnitId: testing
-          ? BannerAd.testAdUnitId
-          : adUnitId.isEmpty
-              ? BannerAd.testAdUnitId
-              : adUnitId.trim(),
-      size: _showSize ?? _setSize ?? AdSize.banner,
-      targetingInfo: targetInfo,
-    );
-  }
+  BannerAd _createAd(
+          {bool? testing, String? adUnitId, AdRequest? targetInfo}) =>
+      BannerAd(
+        size: _showSize ?? _setSize ?? AdSize.banner,
+        adUnitId: testing!
+            ? BannerAd.testAdUnitId
+            : adUnitId!.isEmpty
+                ? BannerAd.testAdUnitId
+                : adUnitId.trim(),
+        listener: listener!.createListener(),
+        request: targetInfo!,
+      );
 }
 
 class FullScreenAd extends MobileAds {
   //
-  factory FullScreenAd({MobileAdListener listener}) =>
+  factory FullScreenAd({required AdMobListener listener}) =>
       _this ??= FullScreenAd._(listener);
+  FullScreenAd._(AdMobListener listener) : super(listener: listener) {
+    listener.closedListener = () {
+      loadAd(show: false);
+    };
+  }
 
-  FullScreenAd._(MobileAdListener listener) : super(listener: listener);
+  static FullScreenAd? _this;
+//  final AdMobListener? interstitialListener;
 
-  static FullScreenAd _this;
+//  final _screenListener = AdMobListener(adEventListeners);
+
+  Set<MobileAdListener> get eventListeners => listener!.eventListeners;
+  // ignore: avoid_setters_without_getters
+  set loadedListener(VoidCallback listener) =>
+      this.listener!.loadedListener = listener;
+  bool removeLoaded(VoidCallback listener) =>
+      this.listener!.removeLoaded(listener);
+  // ignore: avoid_setters_without_getters
+  set failedListener(VoidCallback listener) =>
+      this.listener!.failedListener = listener;
+  bool removeFailed(VoidCallback listener) =>
+      this.listener!.removeFailed(listener);
+  // ignore: avoid_setters_without_getters
+  set clickedListener(VoidCallback listener) =>
+      this.listener!.clickedListener = listener;
+  bool removeClicked(VoidCallback listener) =>
+      this.listener!.removeClicked(listener);
+  // ignore: avoid_setters_without_getters
+  set impressionListener(VoidCallback listener) =>
+      this.listener!.impressionListener = listener;
+  bool removeImpression(VoidCallback listener) =>
+      this.listener!.removeImpression(listener);
+  // ignore: avoid_setters_without_getters
+  set openedListener(VoidCallback listener) =>
+      this.listener!.openedListener = listener;
+  bool removeOpened(VoidCallback listener) =>
+      this.listener!.removeOpened(listener);
+  // ignore: avoid_setters_without_getters
+  set leftAppListener(VoidCallback listener) =>
+      this.listener!.leftAppListener = listener;
+  bool removeLeftApp(VoidCallback listener) =>
+      this.listener!.removeLeftApp(listener);
+  // ignore: avoid_setters_without_getters
+  set closedListener(VoidCallback listener) =>
+      this.listener!.closedListener = listener;
+  bool removeClosed(VoidCallback listener) =>
+      this.listener!.removeClosed(listener);
+  void clearAll() => listener!.clearAll();
 
   /// Set options for the Interstitial Ad
-  // Uses super.set();
   @override
   Future<bool> set({
-    String adUnitId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool nonPersonalizedAds,
-    bool testing,
-    double anchorOffset,
-    double horizontalCenterOffset,
-    AnchorType anchorType,
-    AdErrorListener errorListener,
+    String? adUnitId,
+    AdRequest? targetInfo,
+    List<String>? keywords,
+    String? contentUrl,
+    bool? childDirected,
+    List<String>? testDevices,
+    bool? nonPersonalizedAds,
+    bool? testing,
+    double? anchorOffset,
+    double? horizontalCenterOffset,
+//    AnchorType anchorType,
+    AdErrorListener? errorListener,
   }) {
     super.set(
       adUnitId: adUnitId,
@@ -177,10 +225,10 @@ class FullScreenAd extends MobileAds {
       testing: testing,
       anchorOffset: anchorOffset,
       horizontalCenterOffset: horizontalCenterOffset,
-      anchorType: anchorType,
+//      anchorType: anchorType,
       errorListener: errorListener,
     );
-    return load(
+    return loadAd(
       show: false,
       adUnitId: adUnitId,
       targetInfo: targetInfo,
@@ -191,51 +239,96 @@ class FullScreenAd extends MobileAds {
       testing: testing,
       anchorOffset: anchorOffset,
       horizontalCenterOffset: horizontalCenterOffset,
-      anchorType: anchorType,
+//      anchorType: anchorType,
     );
   }
 
-  _createAd({bool testing, String adUnitId, MobileAdTargetingInfo targetInfo}) {
-    return InterstitialAd(
-      adUnitId: testing
-          ? InterstitialAd.testAdUnitId
-          : adUnitId.isEmpty
-              ? InterstitialAd.testAdUnitId
-              : adUnitId.trim(),
-      targetingInfo: targetInfo,
-    );
-  }
+  @override
+  InterstitialAd _createAd({
+    bool? testing,
+    String? adUnitId,
+    AdRequest? targetInfo,
+  }) =>
+      InterstitialAd(
+        adUnitId: testing!
+            ? InterstitialAd.testAdUnitId
+            : adUnitId!.isEmpty
+                ? InterstitialAd.testAdUnitId
+                : adUnitId.trim(),
+        listener: listener!.createListener(),
+        request: targetInfo!,
+      );
 }
 
 class Native extends MobileAds {
   //
-  factory Native({String factoryId, MobileAdListener listener}) =>
-      _this ??= Native._(factoryId, listener);
+  factory Native({required AdMobListener listener}) =>
+      _this ??= Native._(listener);
+  Native._(AdMobListener listener) : super(listener: listener) {
+    listener.closedListener = () {
+      loadAd(show: false);
+    };
+  }
 
-  Native._(this._factoryId, MobileAdListener listener)
-      : super(listener: listener);
+  static Native? _this;
 
-  static Native _this;
+  Set<MobileAdListener> get eventListeners => listener!.eventListeners;
+  // ignore: avoid_setters_without_getters
+  set loadedListener(VoidCallback listener) =>
+      this.listener!.loadedListener = listener;
+  bool removeLoaded(VoidCallback listener) =>
+      this.listener!.removeLoaded(listener);
+  // ignore: avoid_setters_without_getters
+  set failedListener(VoidCallback listener) =>
+      this.listener!.failedListener = listener;
+  bool removeFailed(VoidCallback listener) =>
+      this.listener!.removeFailed(listener);
+  // ignore: avoid_setters_without_getters
+  set clickedListener(VoidCallback listener) =>
+      this.listener!.clickedListener = listener;
+  bool removeClicked(VoidCallback listener) =>
+      this.listener!.removeClicked(listener);
+  // ignore: avoid_setters_without_getters
+  set impressionListener(VoidCallback listener) =>
+      this.listener!.impressionListener = listener;
+  bool removeImpression(VoidCallback listener) =>
+      this.listener!.removeImpression(listener);
+  // ignore: avoid_setters_without_getters
+  set openedListener(VoidCallback listener) =>
+      this.listener!.openedListener = listener;
+  bool removeOpened(VoidCallback listener) =>
+      this.listener!.removeOpened(listener);
+  // ignore: avoid_setters_without_getters
+  set leftAppListener(VoidCallback listener) =>
+      this.listener!.leftAppListener = listener;
+  bool removeLeftApp(VoidCallback listener) =>
+      this.listener!.removeLeftApp(listener);
+  // ignore: avoid_setters_without_getters
+  set closedListener(VoidCallback listener) =>
+      this.listener!.closedListener = listener;
+  bool removeClosed(VoidCallback listener) =>
+      this.listener!.removeClosed(listener);
+  void clearAll() => listener!.clearAll();
 
   /// An identifier for the factory that creates the Platform view.
-  String _factoryId;
+  String? _factoryId;
 
   /// Set options to the Banner Ad.
   @override
   Future<bool> set({
-    String adUnitId,
-    String factoryId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool nonPersonalizedAds,
-    bool testing,
-    double anchorOffset,
-    double horizontalCenterOffset,
-    AnchorType anchorType,
-    AdErrorListener errorListener,
+    String? adUnitId,
+    String? factoryId,
+    AdRequest? targetInfo,
+    List<String>? keywords,
+    String? contentUrl,
+    bool? childDirected,
+    List<String>? testDevices,
+    bool? nonPersonalizedAds,
+    bool? testing,
+    double? anchorOffset,
+    double? horizontalCenterOffset,
+//    AnchorType anchorType,
+    AdErrorListener? errorListener,
   }) {
     // Override the initial id.
     if (factoryId != null && factoryId.isNotEmpty) {
@@ -252,10 +345,10 @@ class Native extends MobileAds {
       testing: testing,
       anchorOffset: anchorOffset,
       horizontalCenterOffset: horizontalCenterOffset,
-      anchorType: anchorType,
+//      anchorType: anchorType,
       errorListener: errorListener,
     );
-    return load(
+    return loadAd(
       show: false,
       adUnitId: adUnitId,
       targetInfo: targetInfo,
@@ -267,79 +360,181 @@ class Native extends MobileAds {
       testing: testing,
       anchorOffset: anchorOffset,
       horizontalCenterOffset: horizontalCenterOffset,
-      anchorType: anchorType,
+//      anchorType: anchorType,
     );
   }
 
   /// Show the Native Ad.
+  Widget get widget => adWidget();
+
   @override
-  Future<bool> show({
-    String adUnitId,
-    String factoryId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool testing,
-    double anchorOffset,
-    double horizontalCenterOffset,
-    AnchorType anchorType,
-    State state,
-  }) async {
-    // Override the initial id.
-    if (factoryId != null && factoryId.isNotEmpty) {
-      await dispose();
-      _factoryId = factoryId;
-    }
-    return super.show(
-        adUnitId: adUnitId,
-        targetInfo: targetInfo,
-        keywords: keywords,
-        contentUrl: contentUrl,
-        childDirected: childDirected,
-        testDevices: testDevices,
-        testing: testing,
-        anchorOffset: anchorOffset,
-        horizontalCenterOffset: horizontalCenterOffset,
-        anchorType: anchorType,
-        state: state);
+  NativeAd _createAd({
+    bool? testing,
+    String? adUnitId,
+    AdRequest? targetInfo,
+  }) =>
+      NativeAd(
+        adUnitId: testing!
+            ? NativeAd.testAdUnitId
+            : adUnitId!.isEmpty
+                ? NativeAd.testAdUnitId
+                : adUnitId.trim(),
+        factoryId: _factoryId!,
+        listener: listener!.createListener(),
+        request: targetInfo,
+      );
+}
+
+class VideoAd extends MobileAds {
+  //
+  factory VideoAd({required AdMobListener listener}) =>
+      _this ??= VideoAd._(listener);
+  VideoAd._(AdMobListener listener) : super(listener: listener) {
+    listener.closedListener = () {
+      loadAd(show: false);
+    };
   }
 
-  _createAd({bool testing, String adUnitId, MobileAdTargetingInfo targetInfo}) {
-    return NativeAd(
-      adUnitId: testing
-          ? NativeAd.testAdUnitId
-          : adUnitId.isEmpty
-              ? NativeAd.testAdUnitId
-              : adUnitId.trim(),
-      factoryId: _factoryId,
-      targetingInfo: targetInfo,
+  static VideoAd? _this;
+
+  Set<MobileAdListener> get eventListeners => listener!.eventListeners;
+  // ignore: avoid_setters_without_getters
+  set loadedListener(VoidCallback listener) =>
+      this.listener!.loadedListener = listener;
+  bool removeLoaded(VoidCallback listener) =>
+      this.listener!.removeLoaded(listener);
+  // ignore: avoid_setters_without_getters
+  set failedListener(VoidCallback listener) =>
+      this.listener!.failedListener = listener;
+  bool removeFailed(VoidCallback listener) =>
+      this.listener!.removeFailed(listener);
+  // ignore: avoid_setters_without_getters
+  set clickedListener(VoidCallback listener) =>
+      this.listener!.clickedListener = listener;
+  bool removeClicked(VoidCallback listener) =>
+      this.listener!.removeClicked(listener);
+  // ignore: avoid_setters_without_getters
+  set impressionListener(VoidCallback listener) =>
+      this.listener!.impressionListener = listener;
+  bool removeImpression(VoidCallback listener) =>
+      this.listener!.removeImpression(listener);
+  // ignore: avoid_setters_without_getters
+  set openedListener(VoidCallback listener) =>
+      this.listener!.openedListener = listener;
+  bool removeOpened(VoidCallback listener) =>
+      this.listener!.removeOpened(listener);
+  // ignore: avoid_setters_without_getters
+  set leftAppListener(VoidCallback listener) =>
+      this.listener!.leftAppListener = listener;
+  bool removeLeftApp(VoidCallback listener) =>
+      this.listener!.removeLeftApp(listener);
+  // ignore: avoid_setters_without_getters
+  set closedListener(VoidCallback listener) =>
+      this.listener!.closedListener = listener;
+  bool removeClosed(VoidCallback listener) =>
+      this.listener!.removeClosed(listener);
+  void clearAll() => listener!.clearAll();
+
+  /// Set the Video Ad's options.
+  @override
+  Future<bool> set({
+    String? adUnitId,
+    AdRequest? targetInfo,
+    List<String>? keywords,
+    String? contentUrl,
+    bool? childDirected,
+    List<String>? testDevices,
+    bool? nonPersonalizedAds,
+    bool? testing,
+    double? anchorOffset,
+    double? horizontalCenterOffset,
+//    AnchorType anchorType,
+    AdErrorListener? errorListener,
+  }) {
+    super.set(
+      adUnitId: adUnitId,
+      targetInfo: targetInfo,
+      keywords: keywords,
+      contentUrl: contentUrl,
+      childDirected: childDirected,
+      testDevices: testDevices,
+      nonPersonalizedAds: nonPersonalizedAds,
+      testing: testing,
+      anchorOffset: anchorOffset,
+      horizontalCenterOffset: horizontalCenterOffset,
+//      anchorType: anchorType,
+      errorListener: errorListener,
+    );
+    return loadAd(
+      show: false,
+      adUnitId: adUnitId,
+      targetInfo: targetInfo,
+      keywords: keywords,
+      contentUrl: contentUrl,
+      childDirected: childDirected,
+      testDevices: testDevices,
+      testing: testing,
+      anchorOffset: anchorOffset,
+      horizontalCenterOffset: horizontalCenterOffset,
+//      anchorType: anchorType,
     );
   }
+
+  /// Display the Ad
+  @override
+  bool show() => super.show();
+
+  @override
+  RewardedAd _createAd({
+    bool? testing,
+    String? adUnitId,
+    AdRequest? targetInfo,
+  }) =>
+      RewardedAd(
+        adUnitId: testing!
+            ? RewardedAd.testAdUnitId
+            : adUnitId!.isEmpty
+                ? RewardedAd.testAdUnitId
+                : adUnitId.trim(),
+        listener: listener!.createListener(),
+        request: targetInfo,
+      );
 }
 
 abstract class MobileAds extends AdMob {
-  MobileAds({this.listener});
-  MobileAd _ad;
-  final MobileAdListener listener;
-  double _anchorOffset;
-  double _horizontalCenterOffset;
-  AnchorType _anchorType;
+  MobileAds({this.listener}) {
+    listener?.loadedListener = () {
+      loaded = true;
+    };
+  }
 
-  Future<bool> load({
-    @required bool show,
-    String adUnitId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool nonPersonalizedAds,
-    bool testing,
-    double anchorOffset,
-    double horizontalCenterOffset,
-    AnchorType anchorType,
+  final AdMobListener? listener;
+
+  Ad? _ad;
+
+  /// Provide the Ad object
+  Ad? get ad => _ad;
+
+  Key? _key;
+
+  double? _anchorOffset;
+  double? _horizontalCenterOffset;
+//  AnchorType _anchorType;
+  bool loaded = false;
+
+  Future<bool> loadAd({
+    required bool show,
+    String? adUnitId,
+    AdRequest? targetInfo,
+    List<String>? keywords,
+    String? contentUrl,
+    bool? childDirected,
+    List<String>? testDevices,
+    bool? nonPersonalizedAds,
+    bool? testing,
+    double? anchorOffset,
+    double? horizontalCenterOffset,
+//    AnchorType anchorType,
   }) async {
     // Are we testing?
     testing ??= _testing;
@@ -349,67 +544,23 @@ abstract class MobileAds extends AdMob {
       adUnitId = _adUnitId;
     }
 
-    if (targetInfo == null)
-      targetInfo = _targetInfo(
-        keywords: keywords,
-        contentUrl: contentUrl,
-        childDirected: childDirected,
-        testDevices: testDevices,
-        nonPersonalizedAds: nonPersonalizedAds,
-      );
+    targetInfo ??= _targetInfo(
+      keywords: keywords,
+      contentUrl: contentUrl,
+      testDevices: testDevices,
+      nonPersonalizedAds: nonPersonalizedAds,
+    );
 
     // Get rid of the ad if already created.
-    dispose();
+    await dispose();
 
     // Create the ad.
-    _ad =
+    _ad ??=
         _createAd(testing: testing, adUnitId: adUnitId, targetInfo: targetInfo);
 
-    // Assign the ad's listener
-    _ad?.listener = (MobileAdEvent event) {
-      if (show && event == MobileAdEvent.loaded) {
-        try {
-          _ad.show(
-            anchorOffset: anchorOffset ?? _anchorOffset ?? 0.0,
-            horizontalCenterOffset:
-                horizontalCenterOffset ?? _horizontalCenterOffset ?? 0.0,
-            anchorType: anchorType ?? _anchorType ?? Platform.isAndroid
-                ? AnchorType.bottom
-                : AnchorType.top,
-          );
-        } catch (ex) {
-          _setError(ex);
-          return;
-        }
-      }
-
-      if (listener != null) {
-        listener(event);
-      }
-
-      // Dispose and load the ad again.
-      if (event == MobileAdEvent.closed) {
-        dispose().then((_) {
-          load(
-            show: false,
-            adUnitId: adUnitId,
-            targetInfo: targetInfo,
-            keywords: keywords,
-            contentUrl: contentUrl,
-            childDirected: childDirected,
-            testDevices: testDevices,
-            testing: testing,
-            anchorOffset: anchorOffset,
-            horizontalCenterOffset: horizontalCenterOffset,
-            anchorType: anchorType,
-          );
-        });
-      }
-    };
-
-    bool loaded;
     try {
-      loaded = await _ad?.load();
+      await _ad?.load();
+      loaded = true;
     } catch (ex) {
       loaded = false;
       _setError(ex);
@@ -418,30 +569,29 @@ abstract class MobileAds extends AdMob {
     return loaded;
   }
 
-  MobileAd _createAd(
-      {bool testing, String adUnitId, MobileAdTargetingInfo targetInfo});
+  /// todo: Override and implement in a subclass.
+  Ad _createAd({bool? testing, String? adUnitId, AdRequest? targetInfo});
 
   /// Set the MobileAd's options.
   @override
   void set({
-    String adUnitId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool nonPersonalizedAds,
-    bool testing,
-    double anchorOffset,
-    double horizontalCenterOffset,
-    AnchorType anchorType,
-    AdErrorListener errorListener,
+    String? adUnitId,
+    AdRequest? targetInfo,
+    List<String>? keywords,
+    String? contentUrl,
+    bool? childDirected,
+    List<String>? testDevices,
+    bool? nonPersonalizedAds,
+    bool? testing,
+    double? anchorOffset,
+    double? horizontalCenterOffset,
+//    AnchorType anchorType,
+    AdErrorListener? errorListener,
   }) {
     super.set(
       adUnitId: adUnitId,
       keywords: keywords,
       contentUrl: contentUrl,
-      childDirected: childDirected,
       testDevices: testDevices,
       nonPersonalizedAds: nonPersonalizedAds,
       testing: testing,
@@ -449,74 +599,61 @@ abstract class MobileAds extends AdMob {
     );
     _anchorOffset ??= anchorOffset;
     _horizontalCenterOffset ??= horizontalCenterOffset;
-    _anchorType ??= anchorType;
+//    _anchorType ??= anchorType;
   }
 
-  /// Display the MobileAd
-  @override
-  Future<bool> show({
-    String adUnitId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool testing,
-    double anchorOffset,
-    double horizontalCenterOffset,
-    AnchorType anchorType,
-    State state,
-  }) async {
-    if (state != null && !state.mounted) return Future.value(false);
-
-    // Load a new ad if a parameter is passed.
-    if ((adUnitId != null && adUnitId.isNotEmpty) ||
-        targetInfo != null ||
-        (keywords != null && keywords.isNotEmpty) ||
-        (contentUrl != null && contentUrl.isNotEmpty) ||
-        childDirected != null ||
-        (testDevices != null && testDevices.isNotEmpty) ||
-        testing != null ||
-        anchorOffset != null ||
-        horizontalCenterOffset != null ||
-        anchorType != null) {
-      await dispose();
-    }
-
-    // The ad is already loaded. Show it now.
+  /// Explicitly load the Ad
+  Future<bool> load() async {
+    bool loaded = true;
     if (_ad != null) {
       try {
-        bool show = await _ad.show(
-            anchorOffset: anchorOffset ?? _anchorOffset ?? 0.0,
-            horizontalCenterOffset:
-                horizontalCenterOffset ?? _horizontalCenterOffset ?? 0.0,
-            anchorType: anchorType ?? _anchorType ?? AnchorType.bottom);
-        if (show) return show;
+        await _ad?.load();
       } catch (ex) {
-        _ad?.dispose();
-        _ad = null;
-
-        /// try loading it again.
+        loaded = false;
+        _setError(ex);
       }
     }
+    return loaded;
+  }
 
-    // Load the ad and show it when loaded.
-    return load(
-      show: true,
-      adUnitId: adUnitId,
-      targetInfo: targetInfo,
-      keywords: keywords,
-      contentUrl: contentUrl,
-      childDirected: childDirected,
-      testDevices: testDevices,
-      testing: testing,
-      anchorOffset: anchorOffset,
-      horizontalCenterOffset: horizontalCenterOffset,
-      anchorType: anchorType,
-    );
+  /// Determine if the Ad is loaded.
+  Future<bool> isLoaded() async {
+    bool loaded = false;
+    if (_ad != null) {
+      try {
+        loaded = (await _ad?.isLoaded())!;
+      } catch (ex) {
+        loaded = false;
+        _setError(ex);
+      }
+    }
+    return loaded;
+  }
+
+  /// Display as a widget
+  Widget adWidget({Key? key}) {
+    Widget widget;
+    if (_ad == null || _ad is! AdWithView) {
+      widget = Container();
+    } else {
+      widget = AdWidget(key: key ?? _key, ad: _ad as AdWithView);
+    }
+    return widget;
+  }
+
+  /// Display the Ad
+  @override
+  bool show() {
+    final show = loaded && _ad != null && _ad is AdWithoutView;
+    if (show) {
+      loaded = false; // It will have to load again.
+      (_ad as AdWithoutView).show();
+    }
+    return show;
   }
 
   /// Clear the MobileAd reference.
+  @override
   Future<void> dispose() async {
     if (_ad != null) {
       try {
@@ -527,225 +664,22 @@ abstract class MobileAds extends AdMob {
       _ad = null;
     }
   }
-
-  Future<bool> isLoaded() async {
-    bool loaded = false;
-    if (_ad != null) {
-      try {
-        loaded = await _ad?.isLoaded();
-      } catch (ex) {
-        loaded = false;
-        _setError(ex);
-      }
-    }
-    return loaded;
-  }
-}
-
-class VideoAd extends AdMob {
-  //
-  factory VideoAd({RewardedVideoAdListener listener}) =>
-      _this ??= VideoAd._(listener);
-
-  VideoAd._(RewardedVideoAdListener listener) : super() {
-    this.listener = listener;
-  }
-
-  static VideoAd _this;
-
-  RewardedVideoAdListener listener;
-  RewardedVideoAd _rewardedVideoAd;
-
-  /// Set the Video Ad's options.
-  @override
-  Future<bool> set({
-    String adUnitId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool nonPersonalizedAds,
-    bool testing,
-    AdErrorListener errorListener,
-  }) {
-    super.set(
-      adUnitId: adUnitId,
-      keywords: keywords,
-      contentUrl: contentUrl,
-      childDirected: childDirected,
-      testDevices: testDevices,
-      nonPersonalizedAds: nonPersonalizedAds,
-      testing: testing,
-    );
-    // Load into memory.
-    return _loadVideo(
-      show: false,
-      adUnitId: adUnitId,
-      targetInfo: targetInfo,
-      keywords: keywords,
-      contentUrl: contentUrl,
-      childDirected: childDirected,
-      testDevices: testDevices,
-      nonPersonalizedAds: nonPersonalizedAds,
-      testing: testing,
-    );
-  }
-
-  /// Show the Video Ad.
-  @override
-  Future<bool> show({
-    String adUnitId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool nonPersonalizedAds,
-    bool testing,
-    State state,
-  }) async {
-    bool show = false;
-
-    // Don't bother if the app is terminating.
-    if (state != null && !state.mounted) return show;
-
-    // Load a new ad if a parameter is passed.
-    if ((adUnitId != null && adUnitId.isNotEmpty) ||
-        targetInfo != null ||
-        (keywords != null && keywords.isNotEmpty) ||
-        (contentUrl != null && contentUrl.isNotEmpty) ||
-        childDirected != null ||
-        (testDevices != null && testDevices.isNotEmpty) ||
-        testing != null) {
-      _rewardedVideoAd = null;
-    }
-
-    // Show if already loaded.
-    if (_rewardedVideoAd != null) {
-      try {
-        show = await _rewardedVideoAd.show();
-        if (show) return show;
-      } catch (ex) {
-        /// try loading it again.
-        _rewardedVideoAd = null;
-      }
-    }
-
-    // Show the ad when loaded.
-    show = await _loadVideo(
-      show: true,
-      adUnitId: adUnitId,
-      targetInfo: targetInfo,
-      keywords: keywords,
-      contentUrl: contentUrl,
-      childDirected: childDirected,
-      testDevices: testDevices,
-      nonPersonalizedAds: nonPersonalizedAds,
-      testing: testing,
-    );
-
-    return show;
-  }
-
-  Future<bool> _loadVideo({
-    @required bool show,
-    String adUnitId,
-    MobileAdTargetingInfo targetInfo,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool nonPersonalizedAds,
-    bool testing,
-  }) {
-    // Acquire an instance of the Video Ad.
-    _rewardedVideoAd ??= RewardedVideoAd.instance;
-
-    // Assign its listener.
-    _rewardedVideoAd.listener =
-        (RewardedVideoAdEvent event, {String rewardType, int rewardAmount}) {
-      if (show && event == RewardedVideoAdEvent.loaded) {
-        try {
-          _rewardedVideoAd.show();
-        } catch (ex) {
-          _setError(ex);
-          return;
-        }
-      }
-
-      // Call any listeners.
-      if (this.listener != null)
-        this.listener(event,
-            rewardType: rewardType, rewardAmount: rewardAmount);
-
-      // Dispose of the ad and recreate it.
-      if (event == RewardedVideoAdEvent.closed) {
-        dispose();
-        _loadVideo(
-          show: false,
-          adUnitId: adUnitId,
-          targetInfo: targetInfo,
-          keywords: keywords,
-          contentUrl: contentUrl,
-          childDirected: childDirected,
-          testDevices: testDevices,
-          nonPersonalizedAds: nonPersonalizedAds,
-          testing: testing,
-        );
-      }
-    };
-
-    // Assign a valid unit id.
-    if (adUnitId == null || adUnitId.isEmpty || adUnitId.length < 30) {
-      adUnitId = _adUnitId;
-    }
-
-    // Is this a test?
-    testing ??= _testing;
-
-    // If testing, assign a 'test' unit id.
-    String adModId = testing
-        ? RewardedVideoAd.testAdUnitId
-        : adUnitId.isEmpty
-            ? RewardedVideoAd.testAdUnitId
-            : adUnitId.trim();
-
-    if (targetInfo == null)
-      targetInfo = _targetInfo(
-        keywords: keywords,
-        contentUrl: contentUrl,
-        childDirected: childDirected,
-        testDevices: testDevices,
-        nonPersonalizedAds: nonPersonalizedAds,
-      );
-
-    // Load the ad into memory.
-    return _rewardedVideoAd.load(adUnitId: adModId, targetingInfo: targetInfo);
-  }
-
-  /// Clear the video ad plugin.
-  @override
-  void dispose() {
-    _rewardedVideoAd = null;
-  }
 }
 
 abstract class AdMob {
-  String _adUnitId;
-  List<String> _keywords;
-  String _contentUrl; // Can be null
-  bool _childDirected; // Can be null
-  List<String> _testDevices; // Can be null
-  bool _nonPersonalizedAds; // Can be null
-  bool _testing;
+  String? _adUnitId;
+  List<String>? _keywords;
+  String? _contentUrl; // Can be null
+  List<String>? _testDevices; // Can be null
+  bool? _nonPersonalizedAds; // Can be null
+  bool? _testing;
 
-  Exception _ex;
-  Set<AdErrorListener> _adErrorListeners = Set();
+  Exception? _ex;
+  final Set<AdErrorListener> _adErrorListeners = {};
 
   /// Return any exception
-  Exception getError() {
-    Exception ex = _ex;
+  Exception? getError() {
+    final Exception? ex = _ex;
     _ex = null;
     return ex;
   }
@@ -754,17 +688,16 @@ abstract class AdMob {
   bool get inError => _ex != null;
 
   /// Displays an error message if any.
-  String get message => _ex?.toString() ?? "";
+  String get message => _ex?.toString() ?? '';
 
   void set({
-    String adUnitId,
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool nonPersonalizedAds,
-    bool testing,
-    AdErrorListener errorListener,
+    String? adUnitId,
+    List<String>? keywords,
+    String? contentUrl,
+    List<String>? testDevices,
+    bool? nonPersonalizedAds,
+    bool? testing,
+    AdErrorListener? errorListener,
   }) {
     if (adUnitId != null && adUnitId.isNotEmpty && adUnitId.length > 30) {
       _adUnitId ??= adUnitId;
@@ -773,8 +706,6 @@ abstract class AdMob {
     _keywords ??= keywords;
 
     _contentUrl ??= contentUrl;
-
-    _childDirected ??= childDirected;
 
     _testDevices ??= testDevices;
 
@@ -792,19 +723,18 @@ abstract class AdMob {
   void dispose();
 
   /// Return the target audience information
-  MobileAdTargetingInfo _targetInfo({
-    List<String> keywords,
-    String contentUrl,
-    bool childDirected,
-    List<String> testDevices,
-    bool nonPersonalizedAds,
+  AdRequest _targetInfo({
+    List<String>? keywords,
+    String? contentUrl,
+    List<String>? testDevices,
+    bool? nonPersonalizedAds,
   }) {
     // Either might be an 'empty' list.
     keywords ??= _keywords;
 
     // An 'empty' list is not passed.
-    if (keywords.isEmpty ||
-        keywords.every((String s) => s == null || s.isEmpty)) {
+    if (keywords!.isEmpty ||
+        keywords.every((String? s) => s == null || s.isEmpty)) {
       keywords = null;
     }
 
@@ -814,25 +744,20 @@ abstract class AdMob {
       contentUrl = null;
     }
 
-    // If it's true, it has to be passed.
-    if (_childDirected != null && _childDirected)
-      childDirected ??= _childDirected;
-
     // Either might be an 'empty' list.
     testDevices ??= _testDevices;
 
     // An 'empty' list is not passed.
-    if (testDevices.isEmpty ||
-        testDevices.every((String s) => s == null || s.isEmpty)) {
+    if (testDevices!.isEmpty ||
+        testDevices.every((String? s) => s == null || s.isEmpty)) {
       testDevices = null;
     }
 
     nonPersonalizedAds ??= _nonPersonalizedAds;
 
-    return MobileAdTargetingInfo(
+    return AdRequest(
       keywords: keywords,
       contentUrl: contentUrl,
-      childDirected: childDirected,
       testDevices: testDevices,
       nonPersonalizedAds: nonPersonalizedAds,
     );
@@ -849,20 +774,26 @@ abstract class AdMob {
   }
 
   /// The Ad's Error Listener Function.
-  void _adErrorListener(Exception ex) {
-    if (ex == null) return;
-    for (AdErrorListener listener in _adErrorListeners) {
+  void _adErrorListener(Exception? ex) {
+    if (ex == null) {
+      return;
+    }
+    for (final AdErrorListener listener in _adErrorListeners) {
       try {
         listener(ex);
       } catch (e) {
-        print("AdMob: ${e.toString()}");
+        // ignore: avoid_print
+        print('AdMob: ${e.toString()}');
       }
     }
   }
 
   /// Set an Error Event Listener.
-  set errorListener(AdErrorListener listener) {
-    if (listener != null) _adErrorListeners.add(listener);
+  // ignore: avoid_setters_without_getters
+  set errorListener(AdErrorListener? listener) {
+    if (listener != null) {
+      _adErrorListeners.add(listener);
+    }
   }
 
   /// Remove a specific Error Listener.
@@ -871,4 +802,256 @@ abstract class AdMob {
 
   /// Empty any error handlers from memory.
   void clearErrorListeners() => _adErrorListeners.clear();
+}
+
+/// Ads listener
+class AdMobListener {
+  AdMobListener([this._adEventListeners]) {
+    _adEventListeners ??= {};
+  }
+  Set<MobileAdListener>? _adEventListeners;
+
+  Set<MobileAdListener> eventListeners = {};
+
+  /// Listens for when the Ad is loaded in memory.
+  final Set<VoidCallback> _loadedListeners = {};
+  // ignore: avoid_setters_without_getters
+  set loadedListener(VoidCallback listener) => _loadedListeners.add(listener);
+  bool removeLoaded(VoidCallback listener) => _loadedListeners.remove(listener);
+  void _clearLoaded() => _loadedListeners.clear();
+
+  /// Listens for when the Ad fails to display.
+  final Set<VoidCallback> _failedListeners = {};
+  // ignore: avoid_setters_without_getters
+  set failedListener(VoidCallback listener) => _failedListeners.add(listener);
+  bool removeFailed(VoidCallback listener) => _failedListeners.remove(listener);
+  void _clearFailed() => _failedListeners.clear();
+
+  /// Listens for when the Ad is clicked on.
+  final Set<VoidCallback> _clickedListeners = {};
+  // ignore: avoid_setters_without_getters
+  set clickedListener(VoidCallback listener) => _clickedListeners.add(listener);
+  bool removeClicked(VoidCallback listener) =>
+      _clickedListeners.remove(listener);
+  void _clearClicked() => _clickedListeners.clear();
+
+  /// Listens for when the user clicks further on the Ad.
+  final Set<VoidCallback> _impressionListeners = {};
+  // ignore: avoid_setters_without_getters
+  set impressionListener(VoidCallback listener) =>
+      _impressionListeners.add(listener);
+  bool removeImpression(VoidCallback listener) =>
+      _impressionListeners.remove(listener);
+  void _clearImpression() => _impressionListeners.clear();
+
+  /// Listens for when the Ad is opened.
+  final Set<VoidCallback> _openedListeners = {};
+  // ignore: avoid_setters_without_getters
+  set openedListener(VoidCallback listener) => _openedListeners.add(listener);
+  bool removeOpened(VoidCallback listener) => _openedListeners.remove(listener);
+  void _clearOpened() => _openedListeners.clear();
+
+  /// Listens for when the user has left the Ad.
+  final Set<VoidCallback> _leftListeners = {};
+
+  /// todo: Rename to 'exit'
+  // ignore: avoid_setters_without_getters
+  set leftAppListener(VoidCallback listener) => _leftListeners.add(listener);
+  bool removeLeftApp(VoidCallback listener) => _leftListeners.remove(listener);
+  void _clearLeftApp() => _leftListeners.clear();
+
+  /// Listens for when the Ad is closed.
+  final Set<VoidCallback> _closedListeners = {};
+  // ignore: avoid_setters_without_getters
+  set closedListener(VoidCallback listener) => _closedListeners.add(listener);
+  bool removeClosed(VoidCallback listener) => _closedListeners.remove(listener);
+  void _clearClosed() => _closedListeners.clear();
+
+  /// The Ad's Event Listener Function.
+  void eventListener(Ad ad, AdsEvent event) {
+    //
+    for (final listener in _adEventListeners!) {
+      try {
+        listener(event);
+      } catch (ex) {
+        eventError(ex, event: event);
+      }
+    }
+
+    for (final listener in eventListeners) {
+      try {
+        listener(event);
+      } catch (ex) {
+        eventError(ex, event: event);
+      }
+    }
+
+    switch (event) {
+      case AdsEvent.onAdLoaded:
+        for (final listener in _loadedListeners) {
+          try {
+            listener();
+          } catch (ex) {
+            eventError(
+              ex,
+              event: AdsEvent.onAdLoaded,
+            );
+          }
+        }
+
+        break;
+      case AdsEvent.onAdFailedToLoad:
+        for (final listener in _failedListeners) {
+          try {
+            listener();
+          } catch (ex) {
+            eventError(ex, event: AdsEvent.onAdFailedToLoad);
+          }
+        }
+
+        break;
+      case AdsEvent.onNativeAdClicked:
+        for (final listener in _clickedListeners) {
+          try {
+            listener();
+          } catch (ex) {
+            eventError(ex, event: AdsEvent.onNativeAdClicked);
+          }
+        }
+
+        break;
+      case AdsEvent.onNativeAdImpression:
+        for (final listener in _impressionListeners) {
+          try {
+            listener();
+          } catch (ex) {
+            eventError(ex, event: AdsEvent.onNativeAdImpression);
+          }
+        }
+
+        break;
+      case AdsEvent.onAdOpened:
+        for (final listener in _openedListeners) {
+          try {
+            listener();
+          } catch (ex) {
+            eventError(ex, event: AdsEvent.onAdOpened);
+          }
+        }
+
+        break;
+      case AdsEvent.onApplicationExit:
+        for (final listener in _leftListeners) {
+          try {
+            listener();
+          } catch (ex) {
+            eventError(ex, event: AdsEvent.onApplicationExit);
+          }
+        }
+
+        break;
+      case AdsEvent.onAdClosed:
+        for (final listener in _closedListeners) {
+          try {
+            listener();
+          } catch (ex) {
+            eventError(ex, event: AdsEvent.onAdClosed);
+          }
+        }
+
+        break;
+      default:
+    }
+
+    // Notify the developer of any errors.
+    assert(eventErrors.isEmpty, 'Ads: Errors in Ad Events! Refer to logcat.');
+  }
+
+  void clearAll() {
+    _clearLoaded();
+    _clearFailed();
+    _clearClicked();
+    _clearImpression();
+    _clearOpened();
+    _clearLeftApp();
+    _clearClosed();
+  }
+
+  /// Supply a comprehensive Ad Listener to all for multiple listeners assigned.
+  AdListener createListener() => AdListener(
+        onAdLoaded: (Ad ad) => eventListener(ad, AdsEvent.onAdLoaded),
+        onAdFailedToLoad: (Ad ad, LoadAdError error) =>
+            eventListener(ad, AdsEvent.onAdFailedToLoad),
+        onNativeAdClicked: (NativeAd ad) =>
+            eventListener(ad, AdsEvent.onNativeAdClicked),
+        onNativeAdImpression: (NativeAd ad) =>
+            eventListener(ad, AdsEvent.onNativeAdImpression),
+        onAdOpened: (Ad ad) => eventListener(ad, AdsEvent.onAdOpened),
+        onApplicationExit: (Ad ad) =>
+            eventListener(ad, AdsEvent.onApplicationExit),
+        onAdClosed: (Ad ad) => eventListener(ad, AdsEvent.onAdClosed),
+        onRewardedAdUserEarnedReward: (RewardedAd ad, RewardItem reward) =>
+            eventListener(ad, AdsEvent.onRewardedAdUserEarnedReward),
+        onAppEvent: (Ad ad, String name, String data) =>
+            eventListener(ad, AdsEvent.onAppEvent),
+      );
+}
+
+/// The possible events that may occurs working with Ads.
+enum AdsEvent {
+  onAdLoaded,
+  onAdFailedToLoad,
+  onNativeAdClicked,
+  onNativeAdImpression,
+  onAdOpened,
+  onApplicationExit,
+  onAdClosed,
+  onRewardedAdUserEarnedReward,
+  onAppEvent,
+}
+
+List<EventError> eventErrors = [];
+
+List<EventError> getEventErrors() {
+  final List<EventError> list = eventErrors;
+  eventErrors = [];
+  return list;
+}
+
+/// Error Handler for the event listeners.
+void eventError(Object ex, {AdsEvent? event}) {
+  if (ex is! Exception) {
+    ex = Exception(ex.toString());
+  }
+  eventErrors.add(EventError(event, ex));
+  // ignore: avoid_print
+  print('Ads: $event - ${ex.toString()}');
+  // Loop through any error listeners.
+  _adErrorListener(ex);
+}
+
+/// Event Error class
+class EventError {
+  EventError(this.event, this.ex);
+  AdsEvent? event;
+  Exception ex;
+}
+
+Set<MobileAdListener> adEventListeners = {};
+
+Set<AdErrorListener> eventErrorListeners = {};
+
+/// The Ad's Error Listener Function.
+void _adErrorListener(Exception? ex) {
+  if (ex == null) {
+    return;
+  }
+  for (final AdErrorListener listener in eventErrorListeners) {
+    try {
+      listener(ex);
+    } catch (e) {
+      // ignore: avoid_print
+      print('Ads: ${e.toString()}');
+    }
+  }
 }
